+++ src/zimscraperlib/__init__.py
@@ -2,7 +2,11 @@
 # -*- coding: utf-8 -*-
 # vim: ai ts=4 sts=4 et sw=4 nu
 
+import os
+import logging as stdlogging
+
 from .constants import NAME
 from .logging import getLogger
 
-logger = getLogger(NAME)
+debug = os.getenv("ZIMSCRAPERLIB_DEBUG")
+logger = getLogger(NAME, level=stdlogging.DEBUG if debug else stdlogging.INFO)